Fix editing sequences
Editing the description of a sequence confusingly uses the `CkPostEditor`
component. Unfortunately that component (reasonably, given its name)
always assumes that the document it's given is a post. This accidentally
worked fine up until the recent change to coauthor handling, but it now
throws an error when we try to spread `coauthorUserIds` which only
exists on posts, not sequences. This PR changes `CkPostEditor` to be
aware the it's document may be either a post or a sequence. This
probably warrants a more thorough refactor in the future, but this will do
for now.