Fix review prediction market tag for non-admins, and disarm a footgun
The `createdAt` field on posts is only readable by admins; it
corresponds to when a post's first draft revision was created, but in
almost all cases what you want is `postedAt` which is when it was
published.

FooterTagList looked at `createdAt` to decide whether a post is recent,
which doesn't work if you aren't an admin. Change that to `postedAt`.
Also remove `createdAt` from the `PostsBase` fragment, and hunt down
some other relatively-minor bugs of the same type.