Filter out duplicate results in post lists
We previously had a hacky system HideRepeatedPostsContext for hiding
duplicate recommendations in post lists, eg when a post is both curated
and recent, or in the hybrid recommendations both recent and
recommended. That implementation wasn't compatible with suspense, so we
rewrote it, but the rewritten only removed duplicates when they're not
in the same list as each other (but not when the same post appears twice
in the same list, as in the Recombee hybrid recommendations case).

Add client-side duplicate filtering in usePostsList (used by PostsList2)
and RecombeePostsList, to conver that case.