Improve caching behavior of ManifoldProbabilitiesCache
Increased TTL from 30-45s. This now has a randomized component so that
if we're rechecking the cache frequently (eg, if a post with an attached
prediction market is on the front page), then we don't send a bunch of
API requests to Manifold in between when the cache entry expires and
when the updated value is rewritten.

Also, update lastCheckedAt (and check that it changed by enough) before
sending the offsite request. This prevents a situation where, if
checking were to fail (eg if Manifold'd API is down and returns errors),
we would not update the cache timestamp and would send them a lot of
extra API requests that they probably don't want.