Prevent link-circles from wrapping separately from the last word of the link
Links have a small circle after them to indicate hoverability, and to
indicate (via styling) whether the link is visited/read. However if the
link lines up with the right edge, it could wrap separately, leaving it
dangling at the start of a line, which looks awkward.

Fix that by giving the link right padding equal to the size of the link
indicator, then giving the link indicator a zero-width wrapper span so
that it will never line wrap. Since the link indicator is now two nested
spans instead of a single element, it can't be represented by a :after
alone, so it has an explicit span/component, <VisitedIndicator/>.