Fix bug that made vote-buttons on posts (but not comments) sometimes fail to update state
This made it so that when switching a post-vote from up to down or vise
versa, you could get into a state where both vote buttons are colored.
Also made it so that when you vote using the buttons at the top of the
post, it didn't update the buttons at the bottom, and vise versa.