Greatly reduce the number of requests to /analyticsEvent
Switches client-side posts to /analyticsEvent from using fetch to using
sendBeacon, so that they're more likely to succeed when run from
beforeUnloadFired. Increase the flush interval from 1s to 5s, make the
flush timer throttling not trigger on leading edge, and don't bypass the
flush timer during startup. The combined effect of these should be a
greatly reduced number of POST requests to /analyticsEvent (but the same
set of events received), which is better for Vercel hosting costs.