Rewrite background-tab handling to fix analytics and notification-check events being sent for hidden tabs
When you switch away from the site's browser tab, this triggers a
`visibilitychange` event. We intended for that to cause notification
checking and timer anaytics events to stop, but it didn't, because the
event was used to trigger state changes and react rerendering, with the
actual stopping-of-timers linked to useEffects in the rendered
components. This doesn't work because React stops all rendering when the
page is hidden (it might have worked in a past React version, I'm not
sure).

Rewrite AnalyticsPageInitializer, NotificationsEffects, and associated
hooks so that they avoid using React state (using only refs and callback
functions). This makes them able to handle backgrounding correctly.