[WIP] Inline predictions
Adds supprt for inline predictions. When you select text in a post or
comment, the text toolbar has a button to mark the selected text as a
claim and register it as a probability. It then appears in the sidebar
like a react, and other people can also register their probabilities
(similar to the Elicit integration).