Bind wrapped pg-promise functions
Previously we were taking a SqlClient and replacing many of its methods
to add a console log warning if they're slow, but in the process calling
them with an incorrect `this` pointer. This resulted in pg-promise's
configuration being missing, including the connection event handler,
which resulted in the type-parser for vectors not getting configured and
likely also resulted in other subtle breakage.