Silence a false-positive lint from nextjs's eslint plugin
This lint rule would output: "Pages directory cannot be found at
<path>/ckEditor/pages or <path>/ckEditor/src/pages. If using a custom
path, please configure with the `no-html-link-for-pages` rule in your
eslint config file." However it would output it as a console.log, not a
lint failure, for some reason. Disable the rule, since, if it could have
done something helpful, it at least wasn't doing it for lack of
configuration.