Trim URLs before deciding how to hover-preview the
Post https://www.lesswrong.com/posts/BkKHMYp3cmphLZgKd/eating-honey-is-probably-fine-actually contains a link with a leading space in it: ` https://linch.substack.com/p/eating-honey-is-probably-fine-actually`. This caused a surprising symptom: the link was parsed as if it was an onsite link, the `/p/` URL format made it interpreted as a wiki page, and it rendered as a broken redlink. Fix by trimming URLs by deciding how to hover-preview them.