Fix broken dark-mode colors in tables
Introduced in: beb7ac4. The color replacements in darkMode.ts are used
as a writable cache; a refactoring broke this by making it return a new
object each time, including in between writing a value and using it,
which made it so that colors that weren't part of the initial
manually-specified set could not be replaced.