Remove inconsistently-generated nextjs generated file
next-env.d.ts is generated by nextjs on build or run, but is generated
inconsistently (it is generated differently in next build vs next start).
This causes it to often wind up with spurious changes, which is annoying
when switching branches.

Remove it from the repo (it's already in .gitignore) so changes to this
file will be properly ignored.