Prototype fix for error due to aborted fetches on refresh
In Firefox, if you navigate (or initiate some other operation that
involves network fetches) and refresh before those network fetches
finish, then the browser will abort the fetches, this causes a visible
error. This is because refreshing immediately cancels any outstanding
network requests, and JS handles the abort as an error, updating the
page.

This proof-of-concept adds a beforeunload listener and a wrapper around
`fetch()`, which makes it so that once the page starts unloading, any
in-progress fetches remain as unresolved promises.

This is a proof-of-concept and not produciton ready because
`beforeunload` is not the right event; an unload can be aborted, since
other event handlers on the same event can stop the navigation (by
warning about unsaved data and the user clicking cancel). Additionally,
interaction with the back-forward cache is untested.