Fix layout/overlap issues created by making the header variable-height by adding a donation banner to it
Adding a donation banner to <Header> made it taller, causing overlap
issues with some things that positioned themselves based on how tall
they thought it was going to be (notably: top-right buttons on tag
pages, the search results popup, the show/hide effect on the title in
the posts ToC). There was pre-existing tech debt around this owing to
the fact that the header is also variable-height based on breakpoint (it
gets a little thinner on mobile).

Introduce a <HeaderHeightProvider> component which provides a CSS
variable --header-height for other components to use, and makes that
variable vary correctly based on breakpoint and on whether the donation
banner is shown.