Fix sideCommentsCache field (which really shouldn't have been in any fragments) being in a fragment
History-digging suggests that this was added to a fragment so that it
would cause a SQL join which the sideComments resolver would pick up on,
but should have been deleted by permissions before making it to the
client. However, its permissions were permissive, and it was making it
to the client. Remove it from the fragment, and also remove the resolver
bits that were trying to rely on that.