Fix stray commas and duplicate mathjax stylesheets in html highlight for content saved with legacy draftjs
When we truncate a post into an HTML highlight, we remove stylesheets
(in particular, the mathjax stylesheet), truncate, then add them back.
We do this because `truncatise` is a pretty simplistic library and it
would word-count stylesheet contents, and also because in admin-edited
posts a stylesheet can be anywhere in a post but style things that are
earlier in the post so we don't want to truncate it away.

In some early versions of our draftjs mathjax plugin (addressed awhile
ago, but posts saved with the old version still exist), instead of
inserting one copy of the mathjax stylesheet into the post, they would
insert one copy of the mathjax stylesheet _per formula_. The truncation
code would collect all of these, and reinsert them... with comma
separators. This meant that the plain-text meta description for a
particular key post
(https://www.lesswrong.com/posts/EbFABnst8LsidYs5Y/goodhart-taxonomy)
started with a very long run of commas, and also that the post page had
a lot more copies of the mathjax stylesheet than necessary.