Fix server-side debugging
Our scripts/runDevInstance.sh was passing --inspect to turn on
debugging, but this was opening a debugger port to the wrong process.
Move the --inspect to the arguments of next rather than the arguments of
node, which appears to work. (This might have worked under nextjs 15 and
broken under 16, I'm not sure.)