Fix a bug that made draft comments appear in the replies to a comment-permalink for admins
Views (including default views) previously took an optional
ResolverContext, which would be available on the server, but not on the
client (because of mingo). We no longer use mingo, so view functions
only run on the server, so we can make the resolver context required.

The comments default view includes `getDraftSelector` with a default of
`include-my-draft-replies`. When a ResolverContext is provided, that
filters out draft comments, except draft comments by the current user,
as expected. When a ResolverContext is _not_ provided, however, it
includes all draft comments (which then get filtered out for non-admins
by permissions checks further down the line). The view gets used in the
`latestChildren` resolver on `Comments`, which didn't pass a
ResolverContext. Fix by passing a ResolverContext there, then modify
`viewTermsToQuery` and related functions to all treat a context as
required rather than optional, to prevent similar issues.