Fix a double-fetch of post side comments
Due to a quirk/bug in our integration with apollo-client, the query in
`PostBody` to fetch the `sideComments` field led to a double-fetch (one
fetch in SSR, and a second fetch in a request to /graphql). Fix this by
moving the sideComments field to the PostsPage fragment and dropping the
conditional query.